refactor(p2): extract list cards into sibling page (stage 2 of 2)
SearchRecommendationCard + BuiltinFreeSearchCard + BuiltinSourceRow +
SearchServiceEditorTarget + resolveSelectedSearchIndex + AgentSearchEnableCard +
SearchProviderCard + CommonOptions moved from SettingSearchPage.kt (L335-670
of the Stage 1 trimmed coordinator) into new sibling SettingSearchListCards.kt.

Byte-equal: bodies and signatures unchanged.
Visibility widened private -> internal (called by coordinator across files):
SearchRecommendationCard, BuiltinFreeSearchCard, SearchServiceEditorTarget,
resolveSelectedSearchIndex, AgentSearchEnableCard, SearchProviderCard,
CommonOptions. BuiltinSourceRow stays private (only called by
BuiltinFreeSearchCard, both in this sibling).

Coordinator strip: 670 -> 315 (-355 with import cleanup).
Imports trimmed: 18 orphans:
- Stage 2 body-only: Edit01, Delete01, Card, CardDefaults, Surface, Switch,
  TextOverflow, Settings, AutoAIIcon, FormItem, OutlinedNumberInput,
  SearchCommonOptions, Uuid (13).
- Hub-doesn't-use-after-cards-leave: Column, Row, fillMaxWidth, padding,
  MaterialTheme (5).
All confirmed this-refactor orphans (>=1 body ref in 53604a03 baseline,
0 in trimmed coordinator). getValue/setValue (by-delegation) and plus
(operator) kept — used implicitly by hub.

Final coordinator state: 315 lines (down from 1571 baseline, -79.9%).
Just SettingSearchPage hub composable + imports + EOF clean.

Baseline 6/463 unit tests preserved.
3-variant assemble: Debug + Notion + Refactortest all BUILD SUCCESSFUL.
